Renamed styled components, added comments to useFetchPosts
