+++ src/UI/Header.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import * as S from '../styles/HeaderStyled.style';
+import * as S from '../styles/Header.style';
 import Logo from '../styles/Logo.style';
 import { ReactComponent as IconMenu } from './RedditTimerLogo.svg';
 import theme from '../styles/theme';
@@ -7,7 +7,7 @@ import defaultSubreddit from '../sharedVariables';
 
 function Header() {
   return (
-    <S.HeaderStyled>
+    <S.HeaderWrapper>
       <div>
         <Logo to="/" alt="reddit_time_footer_logo" aria-label="go to home page">
           <IconMenu alt="reddit_time_footer_logo" />
@@ -19,7 +19,7 @@ function Header() {
         <S.HeaderLink smooth to="/#how-it-works" theme={theme}>How it works</S.HeaderLink>
         <S.HeaderLink smooth to="/#about" theme={theme}>About</S.HeaderLink>
       </div>
-    </S.HeaderStyled>
+    </S.HeaderWrapper>
   );
 }
