+++ src/components/AmountOfPostsByDayHour.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {
   arrayOf, func, number,
 } from 'prop-types';
-import * as S from '../styles/HeatMapWrapper.style';
+import * as S from '../styles/HeatMap.style';
 import { utcHours } from '../sharedVariables';
 
 // eslint-disable-next-line prefer-arrow-callback
