+++ src/components/SubredditForm.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import * as S from '../styles/SubredditFormWrapper.style';
-import { Heading, ActionButton } from '../styles/HeroSectionStyled.style';
+import * as S from '../styles/SubredditForm.style';
+import { Heading, ActionButton } from '../styles/HeroSection.style';
 import defaultSubreddit from '../sharedVariables';
 
 function SubredditForm() {
